KOHA-1097: Changed logic after requirements.
The value of the 8:th position of the leader was misstakenly
taken into account when deciding the format. This was a fatal
misstake that made movies, music recordings and other recordings
impossible to found in faceted searches. This is now fixed with
a remake of the core code of the plugin.